requirements.txt
pandas==2.3.*
SQLAlchemy==2.0.*
python-dotenv==1.0.*
requests>=2.32,<3